Fix compilation error on armhf with `printf()` and `time_t`
This commit fixes a bug I noticed on the armhf platform.  I guess
`time_t` might only be 32-bits, so to work around this, I cast
it to an `intmax_t` before printing.

+++ src/aprs-is.c
@@ -131,7 +131,7 @@ sendPacket (const char* const restrict server,
 		char* timeoutText = malloc(BUFSIZE);
 		if (timeout > 0)
 		{
-			snprintf(timeoutText, BUFSIZE - 1, " (timeout %ld seconds)", timeout);
+			snprintf(timeoutText, BUFSIZE - 1, " (timeout %jd seconds)", (intmax_t)timeout);
 		}
 		else
 		{
